Also return compliant when value does not exist
If a computer has never been set to have configmgr manage Office 365 updates, this registry key may not exist.  In that case, the remediation script should not run.  If the script doesn't check for a null value, the remediation script will run but error when it tries to update a value that doesn't exist.